requirements.txt
langchain
langchain-core
langchain-google-genai
langchain-community
langchain-chroma
python-dotenv
pypdf